fix: add quotes for fixing values api for reserved keywords (#5937)
Fixes an issue reported by user where if the field name is a reserved
keyword in sql, the values api would error out. By adding quotes around
the field, we specifically indicate that it is a field name and not a
kw, so it works. Also checked for non-kw names, also works.

fixes #5917

Signed-off-by: Yashodhan Joshi <yjdoc2@gmail.com>
Co-authored-by: Yashodhan Joshi <yjdoc2@gmail.com>