fix: histogram visitor (#6042)
Current behaviour is that we expect the first function in a query to be
histogram. If any other function is present (for example `count`) prior
to the histogram then the histogram interval is not set.

We can be more lenient in our parsing and allow other functions to occur
before histogram.

This PR fixes the HistogramVisitor to break the chain only when
histogram function occurs.