fix: Individual level Fine grained RBAC permission for Dashboards (#6173)
Currently, individual level RBAC permission does not work for
dashboards. This pr fixes that. So a role permission like below
<img width="1399" alt="Screenshot 2025-03-04 at 5 27 11 PM"
src="https://github.com/user-attachments/assets/06c2030d-4227-4109-b73f-dc2337151362"
/>

will allow the user to access only the dashboards with `Get` and `All`
permission even if the `Get` permission for the folder is not specified.