fix: check for root user even if org_id is provided in db::user::get (#6207)
Fixes https://github.com/openobserve/openobserve/issues/6171 and
https://github.com/openobserve/openobserve/issues/6206.

In the `db::user::get` function we check for root user only if the
`org_id` is `Some`. This may lead to bugs in many places, if we don't
check for root user first and pass `None` as `org_id` for root user.

Co-authored-by: Hengfei Yang <hengfei.yang@gmail.com>