fix: when pipeline reaches max retries by watch_timeout, skip to next run (#6329)
When a pipeline takes more than the scheduler timeout to complete the
task, the scheduler background job `watch_timeout` increases the
`retries` field by 1 and again updates the scheduled jobs status back to
`waiting` from `processing`. If this happens three times, the `retries`
field value go beyond the max retries and hence is never pulled by
scheduler again.

With this change, whenever we process a derived stream we first check if
the current `retries` field matches with the max retries. If it matches
then we simply make the `retries` `0` and again run it after the
scheduled frequency. The start_time for the query is not changed though,
so there should not be any data loss when quering.