fix: when dashboard owner not present in creation payload, use currently logged in user (#6343)
Currently in the main, when creating a dashboard, we check if the
`owner` field is there in the backend. If it is not present, it returns
Missing Owner error. This PR fixes this by checking first if the `owner`
field is non-empty, otherwise it uses the currently logged in user as
`owner`.