fix: when uploading enrichment table with column name only, do not delete the enrichment table (#6459)
When append_data is false (`?append=false`), and user uploads a
enrichment file with only headers, we replace the whole enrichment table
with a single json record with the headers as keys and empty string as
values. When `append=true` we do nothing.

Also, while uploading an enrichment table, it estimates the probable
json size and checks if that exceeds the total enrichment table size
limit (`ZO_ENRICHMENT_TABLE_LIMIT`) in MB. This pr introduces another
key-value pair in meta table to store the current size of an enrichment
table so that while uploading we can keep a track how much of enrichment
table size is available. This is because, we can't rely on the
stream_stats table which is not realtime and takes some time to update.

It also fixes one bug where if the pipeline name contains `/`, the
scheduler is not able to extract the pipeline id from the scheduled_jobs
module_key.